Update libqcmi_cci.so
Pulled from OOS 3.5.8

For some things to work better, this might be necessary.
An example for that is GPS because the binaries necessary for
GPS use this library as well and since some users have issues
with GPS, I decided to update these blobs and hope that it gets better.

Tester reported that GPS is (still) working for him so that means that
it did not get worse and hence it won't hurt having this.

Change-Id: I832b64539136a1f0fefacc5dd54d6a66e2f44c96